test: pin SSE error-null code, headers, multi-delta frames; move passthrough tests (#17)
